add iteration results of heuristic type 4 & 5 for testcase 10
the results show the type 4 transition disparity option to be a bad heuristic and type 5 to be a good one in this case, however both of them can be considered as bad heuristics in specific cases, for example if there are two automata with 30 and 31 states and also two automata with 3 and 5 states, the state disparity heuristic will select a bad order for joining all the automata (same can happen for transition disparity), in bigger testcases this can lead to a very bad selection order for the join operation. on the other hand both of them can also be considered good heuristics in specific cases. for example if the number of transitions/states in the entire automata list are close to each other (low disparity), selecting these heuristics can lead to optimal results.