add testcase 14 & 15
these testcases are for demonstrating the connectivit
y in the constraint automata, the testcase 14 contains two automaton with 15 and 25 transitions that have 5 alphabets in common, testcase 15 contains two automa
ton exactly the same as the two in 14 but this time they have just two transition labels in common. the first operation results in a CA with 138 transitions while the second one results in a CA with 498 transitions. this shows that it's better to select the automata that have the most transitions labels in common for the join operation first.